haneul-framework: use test instead of build script
One area that contributes very negatively to our compilation times is
haneul-framework crate's build script that builds the system packages. The
build script results in duplicate compilations of a number of crates,
both first and third-party, as well as the actual build time for system
packages. This patch removes the build script in favor of doing the
compilation step in a test function.

Instead of always building the system packages they'll be checked-in to
the repo resulting in no additional build step. Instead a test function
will do the compilation and do a comparison against the checked-in
compiled packages (and docs) vs the newly generated ones, failing if
they differ. The checked-in packages and docs can trivially be updated
by running the test with the `UPDATE` envvar set, e.g.
`UPDATE=1 cargo test -p haneul-framework --test build-system-packages`

Here are some simple timing measurements for running `cargo check` at
the root of the repo on my laptop with and with-out this patch:

  with-patch:
    clean build: 3m 15s
    changing haneul-types: 17.33s
  without-patch:
    clean build:  4m 00s
    changing haneul-types:  1m 42s