[CLI] Fix genesis client yaml file on Windows
## Description 
On Windows, binding the localnet IP address to 0.0.0.0 leads to issues
and failing with a networking error for any CLI command when the
localnet network is the default one. By default, when starting a Haneul
local network using `haneul start` and when there is no default
configuration, the command will generate a genesis including a
`client.yaml` file that contains the localnet IP address set to
`0.0.0.0`. This PR fixes that for the `client.yaml` file and should
close .

## Test plan 

On a Windows box, run the binary and tested that it can connect to the
network.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: 
- [ ] Rust SDK: