[snapshots] Ignore .tmp files during db checkpoint scan
## Description 

During rocksdb snapshotting of the db (for creating rocksdb
checkpoints), we temporarily suffix the toplevel dir with `.tmp`, and
remove it after the snapshotting is complete. There is cleanup code to
ensure that these temp files get deleted if we were to crash
mid-operation, however there may be some rare edge cases where this is
not the case, for example if we crash mid-operation, then start the node
up again without db checkpointing enabled, causing the node to progress
to the next epoch never having executed the cleanup code.

In such cases, we will panic on all subsequent snapshot upload attempts
due to inability to parse the db checkpoint name due to the unexpected
`.tmp` suffix. This handles that edge case by ignoring such files.

## Test plan 

👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: