[docs][site] Fixes small screen nav
## Description 

Fixes broken nav on small screens. Jira
[DOCS-383](https://haneullabs.atlassian.net/browse/DOCS-383) for more info.
Also applies tailwind to some styles, which should eventually be used
exclusively.


[docs/site/src/components/GetStartedLink/index.tsx](https://github.com/GeunhwaJeong/haneul/pull/18546/files#diff-ed1e3e4e3d31a916f33152c76fae5faaa7fad41c38d7436f441e0d977f33ac72):
Removed the logic that controls the display of the Get Started link on
the home page to its own component. This should make changing the
location the link navigates to easier and to change the logic behind
visible/hidden and other updates. Should make updates easier because
only need to change the component reference of the docusaurus component
instead of the related logic.


[docs/site/src/components/ThemeToggle/index.tsx](https://github.com/GeunhwaJeong/haneul/pull/18546/files#diff-3951ee509e2800c54d91ed3424884e59ce6fd2dd8a8d2e71963223728cdad832):
Removed the logic for showing or hiding the default theme switcher from
the default docusaurus component. Moved to its own component to better
support docusaurus updates in the future. Also makes replacing the
default switcher in the future easier.

deleted components from `theme/Navbar`: These components were swizzled
from docusaurus theme but only added a style in some cases and didn't
change anything in others. The Navbar component was ejected with all its
children even though it didn't need to be. The Navbar component (and its
children) are labeled as `not safe` for swizzling. This means that
updates to docusaurus can break any changes made to these components, so
should only be ejected and changed when necessary and requires checking
for updates.


[docs/site/src/theme/Navbar/index.js](https://github.com/GeunhwaJeong/haneul/pull/18546/files#diff-a62ab467816eb68ede0469822b5ed6637ff9af29ab3c5cb3831ccf1b321d5db4):
This component was ejected from Docusaurus 2.x. The changes here are
just what docusaurus updated the component to. And now that I type this,
I realize that I can just delete it so that the default component gets
used instead of the custom one. This means only the default component
`Navbar/Content` is changed by just adding references to the two
components mentioned previously (as opposed to the 5 or 6 components
that were ejected originally).

Styles: Styles were changed where necessary. If a style was updated,
then it was also updated to use tailwind syntax. Two styles were used to
remove the need for swizzling default docusaurus components, which as
mentioned comes at the cost of maintenance). This approach was necessary
because docusaurus reuses a lot of general classes for its components or
creates dynamic ones that are suffixed by the relative path to the
component.
- This style selects the wrapper around the search bar so that the "get
started" button on the home page displays next to the search. Prevents
having to eject Navbar/Search component just to add a class.
    ```
    [class^="navbarSearchContainer"] {
      apply min-[997px]:flex min-[997px]:gap-4
    }  
    ```
- This style selects the theme switch (light/dark). Prevents having to
make an additional update to the Navbar/Content component.
    ```
    button[title^="Switch between dark and light mode"] {
      apply !text-white hover:bg-haneul-blue-primary
    }
    ```

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: