[chore][GraphQL/Limits] Separate QueryLimitChecker extension/factory
## Description

Split up the extension factory from the extension itself, similar to
what we did for `Timeout` before. This avoids the confusion of the
single type being created with defaulted fields to act as the factory
and then creating new versions of itself to act as the extension.

## Test plan

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK: