[chore][GraphQL/Limits] Standardise query_limits_checker file order
## Description

This is a file order change only, with no other meaningful changes. It
standardises the order to:

- Constants
- Types
- Impls
- Trait Impls
- Free functions

## Test plan

CI

## Stack
- 
-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: