[GraphQL/Limits] Reimplement QueryLimitsChecker
## Description

Rewriting query limits checker to land a number of improvements and
fixes:

- Avoid issues with overflows by counting down from a predefined budget,
rather than counting up to the limit and protecting multiplications
using `checked_mul`.

- Improve detection of paginated fields: 
- Previously we treated all connections-related fields as appearing as
many times as the page size (including the field that introduced the
connection, and the `pageInfo` field). This was over-approximated the
output size by a large margin. The new approach counts exactly the
number of nodes in the output: The connection's root field, and any
non-`edges` or `nodes` field will not get multiplied by the page size.
- The checker now also detects connections-related fields even if they
are obscured by fragment or inline fragment spreads.

- Tighter `__schema` query detection: Previously we would skip requests
that started with a `__schema` introspection query. Now it's required to
be the only operation in the request (not just the first).

- Fix metrics collection after limits are hit: Previously, if a limit
was hit, we would not observe validation-related metrics in prometheus.
Now we will always record such metrics, and if a limit has been hit, it
will register as being "at" the limit.

## Test plan

```
haneul-graphql-e2e-tests$ cargo nextest run --features pg_integration -- limits/
```

## Stack

-  
-  
- 
-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: Output node estimation has been made more accurate -- the
estimate should now track the theoretical max number of nodes on the
JSON `data` output.
- [ ] CLI: 
- [ ] Rust SDK: