[DB] increase parallelism per db from 4 to 8, and allow env var override
## Description 

If a DB has a large number of column families, and each of them receives
a large amount of writes, there will be a large number of memtable
flushes and compactions that need to be processed. Only allowing 4
concurrent processing per DB seems not enough, as observed on some
column families stopping writes because of too many memtables pending to
flush. The default is increased to 8.

Allow the value to be overridden by env var, so operators can increase /
decrease the value on their own.

Remove the `set_max_background_jobs()` call since it effect is
overwritten by the subsequent `increase_parallelism()` call.

## Test plan 

Private testnet.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: