[bridge] move add new coin's handling to monitor
## Description 

Today, when receiving an `AddNewToken` event, orchestrator updates the
haneul type tags map and send to the watch channel that action execution
watches. In this PR, we consolidate the handling of this event with the
newly added monitor module. In the new approach, orchestrator sends the
event to monitor, instead of handling it in place. Also the type tag
maps is stored in a `ArcSwap` to allow monitor to change it. Hence we
get rid of the watch channel.

## Test plan 

Added unit tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: