[Consensus] refactored data remover for store pruning
## Description 

This PR refactors the store pruning for the consensus db. The following
have been done:

* renamed component from `EpochDataRemover` to `ConsensusStorePruner` as
it's more accurate and easier to locate
* use the `safe_drop_db` method to ensure more safety/robustness against
deletions
* added node configuration for the consensus db epoch retention and run
interval
* made the component attempt to prune old epoch dbs not only during
epoch change but periodically as well (configurable) to ensure that
there is a retry approach in case of transient failures enhancing the
robustness.
* more testing

## Test plan 

CI/private-testnet

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: