[snapshots] Add metric for number of local db checkpoints
## Description 

Because we run snapshot upload and garbage collection as a background
task but we perform rocksdb checkpointing at end of epoch
unconditionally for configured nodes, any failure in either the upload
or garbage collection path will lead to an accumulation of old db
checkpoints, which will inevitably lead to disk filling and make things
more difficult to debug.

This PR adds a metric to periodically count the number of rocksdb
checkpoints that exist on local disk. Except in rare cases (backfills),
this number should generally be lower than 3, so we can add alerting on
this for early intervention.

## Test plan 

👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: