Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/deepbook-v3@0.2.0

### Minor Changes

-   da3654d: Constants update, manager sdk update

## haneullabs/enoki@0.3.15

### Patch Changes

-   Updated dependencies [1b5cb57]
    -   haneullabs/zklogin@0.7.15

## haneullabs/zklogin@0.7.15

### Patch Changes

-   1b5cb57: Additional check for bigints