[docs] Limit information for object model
## Description 

In its original form, the Object Model document pointed to a line of
code where the ProtocolConfig struct was. Weird, but the code was added
to since the link was made so it no longer pointed to the intended spot.
Worse, the link never showed what the actual limits were, just the param
name and type.

These edits fix the code copy plugin to locate Rust structs so the
current code can be used in the document instead of creating another
soon-to-be outdated link.

Also, adds a component to get the protocol configs for each network so
the values are available from docs.

The Object Model content was also updated to accommodate these features.

Result is bottom of this page:
https://haneul-docs-git-fork-ronny-haneullabs-docs-87-haneul-foundation.vercel.app/concepts/object-model

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: