[chore][GraphQL] Clean-up unused commands
## Description

Remove commands to generate examples, markdown and schema from the main
binary as we do not use them:

- Instead of generating examples, we have hand-crafted examples in our
docs. Removing this code also removes a test that forces regeneration of
a markdown file from these docs (which we also were not using).
- We also never used the output from the `generate-schema` sub-command,
because the schema was always available as a file, or via introspection
commands from the running service.
- Logic for gathering examples to test has been moved into the test
file, to avoid including test-only code in the main library.

## Test plan

CI

## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

## Stack

- 
-  
-  
-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: The GraphQL binary no longer supports generating
examples, or exporting its own schema as these commands have been unused
for some time.
- [ ] CLI: 
- [ ] Rust SDK: