[chore][GraphQL] Declutter schemas
## Description

Remove `draft_target_schema.graphql` and promote
`current_progress_schema.graphql` to be the canonical schema for the
service -- move it to the top-level of the `haneul-graphql-rpc` crate to
make it easier to find. This is to avoid confusion about source of truth
for the GraphQL schema.

Because the TS SDK references the schema at multiple GraphQL versions,
we will need to cherry-pick this change to release branches when it
lands.

## Test plan

CI

## Stack

- 
-  
-  
-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: The schema file has been moved from
`crates/haneul-graphql-rpc/schemas/current_progress_schema.graphql` to
`crates/haneul-graphql-rpc/schema.graphql`.
- [ ] CLI: 
- [ ] Rust SDK: