feat: allow specifying OTEL service name with OTEL_SERVICE_NAME
## Description 

This changes allows specifying the open-telemetry service name to the
`telemetry-subscribers` crate, by setting the `OTEL_SERVICE_NAME`
environment variable. This defaults to haneul-node, so there is no change
in behaviour when unset.

As `OTEL_SERVICE_NAME` is not automatically consulted by the otel rust
implementation, this uses the environment variable to perform the update
in `telemetry-subscribers`.

## Test plan 

The change is minor, the code compiles, and the environment variable is
spelled correctly.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: