[indexer][graphql] Pruner prunes only epoch-partitioned tables
## Description 

Currently, the pruner assumes that all partitioned tables are
partitioned on epoch, which is an issue since `objects_version` is not
partitioned by epoch. Modify the pruner so that it will filter out
non-epoch-partitioned tables, and otherwise do the same thing.

Change `EPOCHS_TO_KEEP` from an env variable to a config, so we can pass
in test values through the transactional test runner, and a prune.move
test to validate that we prune epoch data without any trouble.

## Test plan 

prune.move

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: